fix: move template to correct pages/ directory
The detail-page template was at templates/detail-page/ but the sandbox
sync script discovers templates from templates/pages/. Move to the
correct location so the sandbox preview renders the updated template.

Made-with: Cursor
