--- packages/cli/templates/detail-page/template.doc.mjs
@@ -1,7 +1,0 @@
-/** @type {import('../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'page',
-  name: 'Detail Page',
-  description: 'Order detail page with timeline and line items',
-  isReady: true,
-};
