+++ packages/cli/templates/pages/detail-page/template.doc.mjs
@@ -1,4 +1,4 @@
-/** @type {import('../../../../core/src/docs-types').TemplateDoc} */
+/** @type {import('../../../core/src/docs-types').TemplateDoc} */
 export const doc = {
   type: 'page',
   name: 'Detail Page',
